+++ deposit/src/main/java/com/dev_high/deposit/application/dto/DepositPaymentDto.java
@@ -4,15 +4,16 @@ import com.dev_high.deposit.domain.DepositPayment;
 import com.dev_high.deposit.domain.DepositPaymentMethod;
 import com.dev_high.deposit.domain.DepositPaymentStatus;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public class DepositPaymentDto {
     public record CreateCommand(
             String orderId,
             DepositPaymentMethod method,
-            long amount
+            BigDecimal amount
     ) {
-        public static CreateCommand of(String orderId, DepositPaymentMethod method, long amount) {
+        public static CreateCommand of(String orderId, DepositPaymentMethod method, BigDecimal amount) {
             return new CreateCommand(orderId, method, amount);
         }
     }
@@ -20,9 +21,9 @@ public class DepositPaymentDto {
     public record ConfirmCommand(
             String paymentKey,
             String orderId,
-            Long amount
+            BigDecimal amount
     ) {
-        public static ConfirmCommand of(String paymentKey, String orderId, Long amount) {
+        public static ConfirmCommand of(String paymentKey, String orderId, BigDecimal amount) {
             return new ConfirmCommand(paymentKey, orderId, amount);
         }
     }
@@ -33,7 +34,7 @@ public class DepositPaymentDto {
             String userId,
             String paymentKey,
             String method,
-            long amount,
+            BigDecimal amount,
             OffsetDateTime requestedAt,
             DepositPaymentStatus status,
             String approvalNum,
